move netem3 to netem
