+++ internal/netem/dissector.go
@@ -1,4 +1,4 @@
-package netem3
+package netem
 
 //
 // Protocol dissector
