+++ internal/netem/dpidrop.go
@@ -1,4 +1,4 @@
-package netem3
+package netem
 
 //
 // DPI: rules to drop packets
