+++ internal/netem/dump.go
@@ -1,4 +1,4 @@
-package netem3
+package netem
 
 //
 // Code to dump packets
