+++ internal/netem/unet.go
@@ -1,4 +1,4 @@
-package netem3
+package netem
 
 //
 // UNET: userspace networking.
